+++ zephyr-mcumgr-cli/src/args.rs
@@ -26,6 +26,12 @@ pub struct App {
     #[arg(short, long)]
     pub serial: Option<String>,
 
+    /// Use the given usb serial port as backend
+    ///
+    /// Must contain a regex that matches `vid:pid` or `vid:pid:iface`
+    #[arg(short, long, num_args = 0..=1, default_missing_value = "")]
+    pub usb_serial: Option<String>,
+
     /// Serial port baud rate
     #[arg(short, long, default_value_t = 115200)]
     pub baud: u32,
@@ -39,6 +45,8 @@ pub struct App {
     pub common: CommonArgs,
 
     /// Command group
+    ///
+    /// If missing, run a connection test
     #[command(subcommand)]
-    pub group: Group,
+    pub group: Option<Group>,
 }