+++ zephyr-mcumgr-cli/src/main.rs
@@ -11,7 +11,7 @@ mod formatting;
 use std::time::Duration;
 
 use clap::Parser;
-use zephyr_mcumgr::MCUmgrClient;
+use zephyr_mcumgr::{MCUmgrClient, client::UsbSerialError};
 
 use crate::errors::CliError;
 
@@ -22,16 +22,33 @@ fn cli_main() -> Result<(), CliError> {
 
     let client = if let Some(serial_name) = args.serial {
         let serial = serialport::new(serial_name, args.baud)
+            .timeout(Duration::from_millis(args.timeout))
             .open()
             .map_err(CliError::OpenSerialFailed)?;
         MCUmgrClient::new_from_serial(serial)
+    } else if let Some(identifier) = args.usb_serial {
+        let result = MCUmgrClient::new_from_usb_serial(
+            identifier,
+            args.baud,
+            Duration::from_millis(args.timeout),
+        );
+
+        if let Err(UsbSerialError::IdentifierEmpty { ports }) = &result {
+            println!();
+            if ports.0.is_empty() {
+                println!("No USB serial ports available.");
+            } else {
+                println!("Available USB serial ports:");
+                println!("{}", ports);
+            }
+            println!();
+            std::process::exit(1);
+        }
+
+        result?
     } else {
         return Err(CliError::NoBackendSelected);
     };
-
-    client
-        .set_timeout(Duration::from_millis(args.timeout))
-        .map_err(|e| CliError::SetTimeoutFailed(e.into()))?;
 
     if let Err(e) = client.use_auto_frame_size() {
         log::warn!("Failed to read SMP frame size from device, using slow default");
@@ -39,7 +56,14 @@ fn cli_main() -> Result<(), CliError> {
         log::warn!("Hint: Make sure that `CONFIG_MCUMGR_GRP_OS_MCUMGR_PARAMS` is enabled.");
     }
 
-    groups::run(&client, args.common, args.group)
+    if let Some(group) = args.group {
+        groups::run(&client, args.common, group)?;
+    } else {
+        client.check_connection()?;
+        println!("Device alive and responsive.");
+    }
+
+    Ok(())
 }
 
 fn main() -> miette::Result<()> {