+++ CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - Add support for SMP v1 error's `rsn` field
+- Add `MCUmgrClient::new_from_usb_serial` that connects to a USB VID:PID serial port
+    - Python: `MCUmgrClient::usb_serial`
+    - CLI: add `-u`/`--usb-serial` flag
+        - When no argument specified, list all available ports
+- Add `MCUmgrClient::check_connection` that checks if the device is connected and responding
+    - CLI: run connection test if no group specified
 
 ## [0.4.2] - 2025-12-06